fix: panic on HashTreeRoot/UnmarshalSSZ when bitsize spec value is very large (#220)
* fix: guard bit-size tag conversion against int64 overflow

* fix: avoid overflow in bit-size-> byte-size conversion instead of erroring

* test: shorten bitsize overflow test names, skip on 32-bit

+++ ssztypes/typecache.go
@@ -1937,7 +1937,7 @@ func (tc *TypeCache) buildVectorDescriptor(desc *TypeDescriptor, runtimeType, sc
 			byteLen := sizeHints[0].Size
 			if sizeHints[0].Bits {
 				desc.BitSize = sizeHints[0].Size
-				byteLen = (byteLen + 7) / 8 // ceil up to the next multiple of 8
+				byteLen = int64((uint64(byteLen) + 7) / 8) // ceil up to the next multiple of 8
 			}
 			if byteLen > desc.Len {
 				return sszutils.NewSszErrorf(sszutils.ErrInvalidConstraint, "size hint for vector type is greater than the length of the array (%d > %d)", byteLen, desc.Len)
@@ -1948,7 +1948,7 @@ func (tc *TypeCache) buildVectorDescriptor(desc *TypeDescriptor, runtimeType, sc
 		byteLen := sizeHints[0].Size
 		if sizeHints[0].Bits {
 			desc.BitSize = sizeHints[0].Size
-			byteLen = (byteLen + 7) / 8 // ceil up to the next multiple of 8
+			byteLen = int64((uint64(byteLen) + 7) / 8) // ceil up to the next multiple of 8
 		}
 		desc.Len = byteLen
 	case len(sizeHints) > 0 && sizeHints[0].Expr != "":
